Refactor alert transports to classes
