+++ LibreNMS/Alert/Transport/Mail.php
@@ -1,4 +1,5 @@
-/* Copyright (C) 2015 Daniel Preussker <f0o@librenms.org>
+<?php
+/* Copyright (C) 2014 Daniel Preussker <f0o@devilcode.org>
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -13,27 +14,20 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>. */
 
 /**
- * Clickatell REST-API Transport
- * @author f0o <f0o@librenms.org>
- * @copyright 2015 f0o, LibreNMS
+ * Mail Transport
+ * @author f0o <f0o@devilcode.org>
+ * @copyright 2014 f0o, LibreNMS
  * @license GPL
  * @package LibreNMS
  * @subpackage Alerts
  */
+namespace LibreNMS\Alert\Transport;
 
-$url  = 'https://platform.clickatell.com/messages/http/send?apiKey='.$opts['token'].'&to='.implode(',',$opts['to']).'&content='.urlencode($obj['title']);
+use LibreNMS\Interfaces\Alert\Transport;
 
-$curl = curl_init($url);
-set_curl_proxy($curl);
-curl_setopt($curl, CURLOPT_CUSTOMREQUEST, "GET");
-curl_setopt($curl, CURLOPT_RETURNTRANSFER, true);
-
-$ret  = curl_exec($curl);
-$code = curl_getinfo($curl, CURLINFO_HTTP_CODE);
-if( $code > 200 ) {
-    if( $debug ) {
-        var_dump($ret);
+class Mail implements Transport{
+    public function call($obj, $opts) {
+        global $config;
+        return send_mail($obj['contacts'], $obj['title'], $obj['msg'], ($config['email_html'] == 'true') ? true : false);
     }
-    return false;
 }
-return true;
