+++ includes/alerts.inc.php
@@ -843,13 +843,14 @@ function ExtTransports($obj)
         if (is_array($opts)) {
             $opts = array_filter($opts);
         }
-        if (($opts === true || !empty($opts)) && $opts != false && file_exists($config['install_dir'].'/includes/alerts/transport.'.$transport.'.php')) {
+        $class  = 'LibreNMS\\Alert\\Transport\\' . ucfirst($transport);
+        if (($opts === true || !empty($opts)) && $opts != false && class_exists($class)) {
             $obj['transport'] = $transport;
             $msg        = FormatAlertTpl($obj);
             $obj['msg'] = $msg;
             echo $transport.' => ';
-            eval('$tmp = function($obj,$opts) { global $config; '.file_get_contents($config['install_dir'].'/includes/alerts/transport.'.$transport.'.php').' return false; };');
-            $tmp = $tmp($obj, $opts);
+            $instance = new $class;
+            $tmp = $instance->call($obj, $opts);
             $prefix = array( 0=>"recovery", 1=>$obj['severity']." alert", 2=>"acknowledgment" );
             $prefix[3] = &$prefix[0];
             $prefix[4] = &$prefix[0];
